backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
JustWatch==0.5.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
RapidFuzz==3.13.0
requests==2.32.4
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
yt-dlp==2025.6.9
